+++ uefi-services/CHANGELOG.md
@@ -1,8 +1,8 @@
 # uefi-services - [Unreleased]
 
 ## Changed
-- The implicit `qemu-exit` crate feature has been removed. (Note that this is
-  different from the `qemu` crate feature, which is unchanged.)
+- `uefi-services` is deprecated and should be removed. All functionality was
+  moved to `uefi::helpers::init()`
 
 # uefi-services - 0.23.0 (2023-11-12)